requirements.txt
# FastAPI & API
fastapi==0.120.4
uvicorn[standard]==0.38.0
python-multipart==0.0.20
aiofiles==25.1.0
requests==2.32.5
redis==7.0.1
SQLAlchemy==2.0.44
aiosqlite==0.21.0
Jinja2==3.1.6

# Научные и ML-библиотеки
numpy==2.3.4
scipy==1.16.3
scikit-learn==1.7.2
tqdm==4.67.1
pandas==2.3.3
Pillow==12.0.0

# ML / векторные модели
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.12.0

# Утилиты
python-dotenv==1.2.1
PyYAML==6.0.3
typing-extensions==4.15.0

